test: refactor directory layout.
